+++ src/main/java/nextstep/repository/ThemeRepository.java
@@ -1,6 +1,6 @@
 package nextstep.repository;
 
-import nextstep.Theme;
+import nextstep.domain.Theme;
 
 import java.util.List;
